requirements.txt
pyarrow
pandas
numpy
scipy
matplotlib
qablet_contracts
qablet-basic
pytest
nbmake
ruff
nb-clean
line_profiler
finmc